# services/exam-service/docs/swagger/swagger.yaml
basePath: /
definitions:
  DifficultyLevel:
    enum:
    - Easy
    - Medium
    - Hard
    type: string
    x-enum-varnames:
    - DifficultyEasy
    - DifficultyMedium
    - DifficultyHard
  Exam:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      description:
        type: string
      durationMinutes:
        type: integer
      enterpriseId:
        type: string
      id:
        type: string
      invitationMethod:
        description: Email, Link, Token
        type: string
      maxParticipants:
        type: integer
      negativeMarking:
        type: boolean
      passingScorePercent:
        type: number
      questions:
        description: Relational data
        items:
          $ref: '#/definitions/ExamQuestion'
        type: array
      randomizationRules:
        items:
          $ref: '#/definitions/ExamRandomizationRule'
        type: array
      scheduledEnd:
        type: string
      scheduledStart:
        type: string
      settings:
        additionalProperties: true
        type: object
      status:
        $ref: '#/definitions/ExamStatus'
      templateSourceId:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  ExamQuestion:
    properties:
      examId:
        type: string
      id:
        type: string
      orderIndex:
        type: integer
      pointsOverride:
        type: integer
      question:
        allOf:
        - $ref: '#/definitions/Question'
        description: populated if joined
      questionId:
        type: string
    type: object
  ExamRandomizationRule:
    properties:
      difficulty:
        $ref: '#/definitions/DifficultyLevel'
      examId:
        type: string
      id:
        type: string
      questionCount:
        type: integer
      topic:
        type: string
    type: object
  ExamStatus:
    enum:
    - Draft
    - Scheduled
    - Active
    - Closed
    - Archived
    type: string
    x-enum-varnames:
    - ExamDraft
    - ExamScheduled
    - ExamActive
    - ExamClosed
    - ExamArchived
...
  CloneExamRequest:
    properties:
      title:
        type: string
    required:
    - title
    type: object
  ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ExamRuleRequest:
    properties:
      difficulty:
        $ref: '#/definitions/DifficultyLevel'
      questionCount:
        type: integer
      topic:
        type: string
    required:
    - questionCount
    type: object
...
info:
  contact:
    name: Veritas Platform Team
  description: Exam authoring, scheduling, publishing, and question bank management
    service.
  title: Veritas Exam Service API
  version: "1.0"
paths:
  /exams:
    get:
      description: List all exams for the caller enterprise.
      parameters:
      - description: Enterprise ID (UUID)
        in: header
        name: X-Enterprise-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Exam'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: List exams
      tags:
      - exam
    post:
      consumes:
      - application/json
      description: Create one exam under the caller enterprise.
      parameters:
      - description: Enterprise ID (UUID)
        in: header
        name: X-Enterprise-ID
        required: true
        type: string
      - description: Actor user ID (UUID)
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Exam payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/Exam'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Exam'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create exam
      tags:
      - exam
  /exams/{examId}:
    delete:
      description: Delete one exam by ID.
      parameters:
      - description: Enterprise ID (UUID)
        in: header
        name: X-Enterprise-ID
        required: true
        type: string
      - description: Exam ID (UUID)
        in: path
        name: examId
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Delete exam
      tags:
      - exam
    get:
      description: Get one exam for the caller enterprise.
      parameters:
      - description: Enterprise ID (UUID)
        in: header
        name: X-Enterprise-ID
        required: true
        type: string
      - description: Exam ID (UUID)
        in: path
        name: examId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Exam'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get exam
      tags:
      - exam
    patch:
      consumes:
      - application/json
      description: Update exam fields by ID.
      parameters:
      - description: Enterprise ID (UUID)
        in: header
        name: X-Enterprise-ID
        required: true
        type: string
      - description: Actor user ID (UUID)
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Exam ID (UUID)
        in: path
        name: examId
        required: true
        type: string
      - description: Exam payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/Exam'
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Update exam
      tags:
      - exam
  /exams/{examId}/clone:
    post:
      consumes:
      - application/json
      description: Clone exam content into a new exam with provided title.
      parameters:
      - description: Enterprise ID (UUID)
        in: header
        name: X-Enterprise-ID
        required: true
        type: string
      - description: Actor user ID (UUID)
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Source Exam ID (UUID)
        in: path
        name: examId
        required: true
        type: string
      - description: Clone payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CloneExamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Exam'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Clone exam
      tags:
      - exam
  /exams/{examId}/close:
    post:
      description: Close an active exam.
      parameters:
      - description: Enterprise ID (UUID)
        in: header
        name: X-Enterprise-ID
        required: true
        type: string
      - description: Exam ID (UUID)
        in: path
        name: examId
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Close exam
      tags:
      - exam
  /exams/{examId}/publish:
    post:
      description: Publish exam after validation checks.
      parameters:
      - description: Enterprise ID (UUID)
        in: header
        name: X-Enterprise-ID
        required: true
        type: string
      - description: Exam ID (UUID)
        in: path
        name: examId
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Publish exam
      tags:
      - exam
...
  /exams/{examId}/rules:
    post:
      consumes:
      - application/json
      description: Add rule for selecting random questions by topic/difficulty.
      parameters:
      - description: Enterprise ID (UUID)
        in: header
        name: X-Enterprise-ID
        required: true
        type: string
      - description: Exam ID (UUID)
        in: path
        name: examId
        required: true
        type: string
      - description: Rule payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ExamRuleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ExamRandomizationRule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Add randomization rule
      tags:
      - exam
...
schemes:
- http
- https
swagger: "2.0"
...